+++ src/components/ProfileImage.tsx
@@ -2,7 +2,7 @@ import { ProfilePopover } from './ProfilePopover';
 import { useWallet } from '@solana/wallet-adapter-react';
 import { getPFPFromPublicKey } from '@/modules/utils/image';
 import { useConnectedWalletProfile } from 'src/views/_global/ConnectedWalletProfileProvider';
-import PopoverWidget from './PopoverWidget';
+import Popover from './Popover';
 
 export const ProfileImage = () => {
   const { connectedProfile } = useConnectedWalletProfile();
@@ -14,11 +14,11 @@ export const ProfileImage = () => {
 
   return (
     <>
-      <PopoverWidget content={<ProfilePopover />}>
+      <Popover content={<ProfilePopover />}>
         <button className="flex items-center justify-center overflow-hidden rounded-full shadow-lg shadow-black ring-4 ring-gray-900 transition-transform  hover:scale-125">
           <img width={44} height={44} src={profilePictureUrl} className={` `} alt="Profile Image" />
         </button>
-      </PopoverWidget>
+      </Popover>
     </>
   );
 };
